dose admin

+++ 106_A2_M1/Model/Vaccine.cs
@@ -9,7 +9,7 @@ namespace _106_A2_M1.Model
     public class Vaccine
     {
         public string dose_id { get; set; }
-        public int date_administered { get; set; }
+        public long date_administered { get; set; }
         public string brand { get; set; }
         public string location { get; set; }
     }
